TableStorage/requirements.txt
python-dotenv
azure-identity
azure-data-tables
azure-core